카테고리 서비스 테스트추가, 주석 수정 및 로직 개선

+++ src/main/java/backend/graduationprojectspring/service/impl/CategoryServiceImpl.java
@@ -19,12 +19,6 @@ public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepo categoryRepo;
     private final CategoryQueryRepo categoryQueryRepo;
 
-    /**
-     * 카테고리 데이터 베이스에 저장
-     * @param name 저장할 카테고리 이름
-     * @return 저장된 카테고리 반환
-     * @throws DuplicateException 같은 이름으로 존재하는 카테고리가 있을 경우
-     */
     @Override
     public Category create(String name){
         if(categoryRepo.existsByName(name)){
@@ -34,35 +28,18 @@ public class CategoryServiceImpl implements CategoryService {
         return categoryRepo.save(category);
     }
 
-    /**
-     * 카테고리 페이지 조회(카테고리 이름으로 정렬됨)<br>
-     * @param page 현재 보여줄 페이지 위치
-     * @param size 얼마만큼 보여줄지 크기
-     * @return 조회된 category List 반환
-     */
     @Override
     @Transactional(readOnly = true)
     public List<Category> paging(int page, int size){
         return categoryQueryRepo.paging(page, size);
     }
 
-    /**
-     * 카테고리 전체 개수 반환
-     * @return 카테고리 전체 수
-     */
     @Override
     @Transactional(readOnly = true)
     public long totalCount(){
         return categoryRepo.count();
     }
 
-    /**
-     * 카테고리 이름 수정
-     * @param id 이름을 수정할 카테고리의 id
-     * @param updateName 수정할 이름
-     * @throws NotExistsException id에 해당하는 카테고리가 없으면 예외 발생
-     * @throws DuplicateException 수정할 이름으로 존재하는 카테고리가 이미 존재하면 발생
-     */
     @Override
     public void updateName(Long id, String updateName){
         Category findCategory = categoryRepo.findById(id)
@@ -73,20 +50,14 @@ public class CategoryServiceImpl implements CategoryService {
         findCategory.updateName(updateName);
     }
 
-    /**
-     * 카테고리 삭제
-     * @param id 삭제할 카테고리의 id
-     * @throws IllegalArgumentException id가 null이 들어올 경우 예외 반환
-     */
     @Override
     public void delete(Long id){
+        if(id == null){
+            throw new IllegalArgumentException("id가 null이 올 수 없습니다.");
+        }
         categoryRepo.deleteById(id);
     }
 
-    /**
-     * 모든 카테고리 반환
-     * @return 모든 카테고리 리스트 반환
-     */
     @Override
     @Transactional(readOnly = true)
     public List<Category> findAll(){
